+++ Cryptography/RSA/RSA.cs
@@ -5,7 +5,7 @@ namespace Cryptography.RSA;
 
 public class RSA
 {
-    private static readonly ulong[] _primes;
+    private static readonly UInt128[] _primes;
 
     private static readonly Random rand;
 
@@ -15,9 +15,9 @@ public class RSA
         rand = new Random();
     }
 
-    private static ulong[] GetPrimesFromFile()
+    private static UInt128[] GetPrimesFromFile()
     {
-        var primes = new List<ulong>();
+        var primes = new List<UInt128>();
 
         if (!File.Exists(Path.Combine(Directory.GetCurrentDirectory(), "Primes.txt")))
         {
@@ -29,7 +29,7 @@ public class RSA
 
         using (var file = new StreamReader(File.OpenRead(Path.Combine(Directory.GetCurrentDirectory(), "Primes.txt"))))
         {
-            while (!file.EndOfStream) primes.Add(ulong.Parse(file.ReadLine()));
+            while (!file.EndOfStream) primes.Add(UInt128.Parse(file.ReadLine()));
         }
 
         return primes.ToArray();
@@ -45,14 +45,14 @@ public class RSA
         return (results.y - b / a * results.x, results.x, results.gcd);
     }
 
-    private static ulong EulerFunctionForCompositeNumber(ulong a, ulong b)
+    private static UInt128 EulerFunctionForCompositeNumber(UInt128 a, UInt128 b)
     {
         return (a - 1) * (b - 1);
     }
 
-    private static ulong[] GetRandomPrimes(ulong maxValue, int amount)
+    private static UInt128[] GetRandomPrimes(UInt128 maxValue, int amount)
     {
-        var selectedPrimes = new ulong[amount];
+        var selectedPrimes = new UInt128[amount];
 
         for (var i = 0; i < amount; i++)
         {
@@ -69,7 +69,7 @@ public class RSA
         return selectedPrimes;
     }
 
-    private static BigInteger Power(ulong x, ulong n)
+    private static BigInteger Power(UInt128 x, UInt128 n)
     {
         BigInteger result = 1;
         BigInteger bufX = x;
@@ -88,13 +88,13 @@ public class RSA
         return result;
     }
 
-    private static ulong ModInversion(ulong value, ulong modulo)
+    private static UInt128 ModInversion(UInt128 value, UInt128 modulo)
     {
         var res = EuclidExtendedAlgorithhm(value, modulo);
-        return res.gcd > 1 ? 0 : (ulong)(res.x % modulo + modulo) % modulo;
+        return res.gcd > 1 ? 0 : (UInt128)(res.x % modulo + modulo) % modulo;
     }
 
-    private static (ulong e, ulong d, ulong r) GetKeys()
+    private static (UInt128 e, UInt128 d, UInt128 r) GetKeys()
     {
         var primes = GetRandomPrimes(ulong.MaxValue, 2);
 
@@ -113,26 +113,26 @@ public class RSA
         return (e, d, r);
     }
 
-    private static ulong[] Encode(ulong[] text, ulong key, ulong r)
+    private static UInt128[] Encode(UInt128[] text, UInt128 key, UInt128 r)
     {
         Parallel.For(0, text.Length, (i) =>
         {
-            text[i] = (ulong)(Power(text[i], key) % r);
+            text[i] = (UInt128)(Power(text[i], key) % r);
         });
 
         return text;
     }
 
-    public static (ulong[] encryptedText, ulong key, ulong r) Encrypt(string text)
+    public static (UInt128[] encryptedText, UInt128 key, UInt128 r) Encrypt(string text)
     {
         var keys = GetKeys();
 
-        var encryptedStr = Encode(text.Select(x => (ulong)x + 2).ToArray(), keys.e, keys.r);
+        var encryptedStr = Encode(text.Select(x => (UInt128)x + 2).ToArray(), keys.e, keys.r);
 
         return (encryptedStr, keys.d, keys.r);
     }
 
-    public static string Decrypt(ulong[] text, ulong key, ulong r)
+    public static string Decrypt(UInt128[] text, UInt128 key, UInt128 r)
     {
         var sb = new StringBuilder(text.Length);
         sb.Append(Encode(text, key, r).Select(x => (char)(x - 2)).ToArray());
